+++ khoros/objects/messages.py
@@ -7,7 +7,7 @@
                     node_id='support-tkb')``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     11 Oct 2021
+:Modified Date:     10 Jan 2022
 """
 
 import json
@@ -45,6 +45,9 @@ def create(khoros_object, subject=None, body=None, node=None, node_id=None, node
            return_http_code=False, return_status=None, return_error_messages=None, split_errors=False,
            proxy_user_object=None):
     """This function creates a new message within a given node.
+
+    .. versionchanged:: 4.5.0
+       The Content-Type header is now explicitly defined as ``application/json`` when handling non-multipart requests.
 
     .. versionchanged:: 4.4.0
        Introduced the ``proxy_user_object`` parameter to allow messages to be created on behalf of other users.
@@ -169,8 +172,9 @@ def create(khoros_object, subject=None, body=None, node=None, node_id=None, node
     multipart = True if attachment_file_paths else False
     if multipart:
         payload = attachments.construct_multipart_payload(payload, attachment_file_paths)
+    content_type = 'application/json' if not multipart else None
     response = api.post_request_with_retries(api_url, payload, khoros_object=khoros_object, multipart=multipart,
-                                             proxy_user_object=proxy_user_object)
+                                             content_type=content_type, proxy_user_object=proxy_user_object)
     return api.deliver_v2_results(response, full_response, return_id, return_url, return_api_url, return_http_code,
                                   return_status, return_error_messages, split_errors, khoros_object)
 
@@ -513,6 +517,37 @@ def update(khoros_object, msg_id=None, msg_url=None, subject=None, body=None, no
                                   return_status, return_error_messages, split_errors, khoros_object)
 
 
+def get_metadata(khoros_object, msg_id, metadata_key):
+    """This function retrieves the value for a specific metadata key associated with a given message.
+
+    .. versionadded:: 4.5.0
+
+    :param khoros_object: The core :py:class:`khoros.Khoros` object
+    :type khoros_object: class[khoros.Khoros]
+    :param msg_id: The ID of the message for which the metadata will be retrieved
+    :type msg_id: str, int
+    :param metadata_key: The metadata key for which the value will be retrieved
+    :type metadata_key: str
+    :returns: The metadata value
+    :raises: :py:exc:`khoros.errors.exceptions.MissingRequiredDataError',
+             :py:exc:`khoros.errors.exceptions.InvalidMetadataError`,
+             :py:exc:`khoros.errors.exceptions.GETRequestError`
+    """
+    if not msg_id or not metadata_key:
+        raise errors.exceptions.MissingRequiredDataError('A message ID and a metadata key are required')
+    uri = f'/messages/id/{msg_id}/metadata/key/{metadata_key}'
+    response = khoros_object.v1.get(uri)
+    if not response.get('response'):
+        raise errors.exceptions.GETRequestError('The GET request to retrieve the message metadata was not successful')
+    if response['response'].get('status') == 'error':
+        if response['response']['error'].get('message'):
+            raise errors.exceptions.InvalidMetadataError(response['response']['error'].get('message'))
+        else:
+            raise errors.exceptions.InvalidMetadataError()
+    metadata_value = response['response']['value'].get('$')
+    return metadata_value
+
+
 def _verify_message_id(_msg_id, _msg_url):
     """This function verifies that a message ID has been defined or can be using the message URL.
 
@@ -527,7 +562,7 @@ def _verify_message_id(_msg_id, _msg_url):
              :py:exc:`errors.exceptions.MessageTypeNotFoundError`
     """
     if not any((_msg_id, _msg_url)):
-        raise errors.exceptions.MissingRequiredDataError("A message ID or URL must be defined when updating messages")
+        raise errors.exceptions.MissingRequiredDataError('A message ID or URL must be defined when updating messages')
     elif not _msg_id:
         _msg_id = get_id_from_url(_msg_url)
     return _msg_id